% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admin}
\alias{admin}
\title{admin boundary dataset to be used as an example}
\format{
A spatial polygon data frame with 12 rows and 13 variables:
\describe{
  \item{\code{OBJECTID_1}}{character COLUMN_DESCRIPTION}
  \item{\code{OBJECTID}}{character COLUMN_DESCRIPTION}
  \item{\code{GEOCODE11}}{double COLUMN_DESCRIPTION}
  \item{\code{LANDTYPE}}{character COLUMN_DESCRIPTION}
  \item{\code{Division}}{character COLUMN_DESCRIPTION}
  \item{\code{District}}{character COLUMN_DESCRIPTION}
  \item{\code{Upazila}}{character COLUMN_DESCRIPTION}
  \item{\code{Union}}{character COLUMN_DESCRIPTION}
  \item{\code{UpzCode}}{double COLUMN_DESCRIPTION}
  \item{\code{GEOCODE15}}{character COLUMN_DESCRIPTION}
  \item{\code{Shape_Leng}}{double COLUMN_DESCRIPTION}
  \item{\code{Shape_Le_1}}{double COLUMN_DESCRIPTION}
  \item{\code{Shape_Area}}{double COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
admin
}
\description{
This dataset contains the boudaries for the Unions in both Ukhia and Teknaf.
}
\keyword{internal}
