% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_a.R
\name{assessmentDailyValidSurveys}
\alias{assessmentDailyValidSurveys}
\title{Daily number of survey per consent status}
\usage{
assessmentDailyValidSurveys(
  ds = NULL,
  surveyDate = NULL,
  dateFormat = NULL,
  surveyConsent = NULL
)
}
\arguments{
\item{ds}{dataset containing the survey (from kobo): data.frame}

\item{surveyDate}{name of the field in the dataset where the date of the survey is stored: string}

\item{dateFormat}{format used for the date: string ('\%m/\%d/\%Y')}

\item{surveyConsent}{name of the field in the dataset where the survey consent is stored: string}
}
\value{
dst      same dataset as the inputed one but with survey marked for deletion if errors are found and delete=TRUE (or NULL)

ret_log  list of the errors found (or NULL)

var      a list of value (or NULL)

graph    graphical representation of the results (or NULL)
}
\description{
This function display the number of surveys conducted per day per constent status.
}
\examples{
{
ds <- HighFrequencyChecks::sample_dataset
surveyDate <- "survey_date"
dateFormat <- "\%m/\%d/\%Y"
surveyConsent <- "survey_consent"

list[dst,ret_log,var,graph] <- assessmentDailyValidSurveys(ds=ds,
                                                           surveyDate=surveyDate,
                                                           dateFormat=dateFormat,
                                                           surveyConsent=surveyConsent)
head(ret_log,10)
print(graph)
}
}
\author{
Yannick Pascaud
}
