% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{assessmentDuration}
\alias{assessmentDuration}
\title{Compute the average and total time for the surveys}
\usage{
assessmentDuration(ds = NULL, dates = NULL)
}
\arguments{
\item{ds}{dataset as a data.frame object}

\item{dates}{fields as a list of string where the survey start and end date is stored (c('start_date','end_date'))}
}
\value{
avg  average time per survey

tot total time
}
\description{
This function compute the average and total time for the surveys
Warning: If there are uncorrected mistakes in the survey dates, it can lead to have the length of the survey in seconds and this check will not performed well
}
\examples{
{
ds <- HighFrequencyChecks::sample_dataset
dates <- c("survey_start","end_survey")

assessmentDuration(ds, dates)
}
}
\author{
Yannick Pascaud
}
