% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{assessmentProductivity}
\alias{assessmentProductivity}
\title{Summary of daily average productivity}
\usage{
assessmentProductivity(
  ds = NULL,
  surveydate = NULL,
  dateformat = NULL,
  survey_consent = NULL
)
}
\arguments{
\item{ds}{dataset as a data.frame object}

\item{surveydate}{name as a string of the field in the dataset where the date of the survey is stored}

\item{dateformat}{format as a string used for the date ('\%m/\%d/\%Y')}

\item{survey_consent}{name as a string of the field in the dataset where the survey consent is stored}
}
\value{
logf  the report
}
\description{
This function display the number of surveys conducted per day.
}
\examples{
\dontrun{
df <- HighFrequencyChecks::sample_dataset
sdte <- "survey_date"
dtf <- "\%m/\%d/\%Y"
sc <- "survey_consent"

log <- assessmentProductivity(df, sdte, dtf, sc)
head(log,10)
}
}
\author{
Yannick Pascaud
}
