% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{assessmentProductivityGraphical}
\alias{assessmentProductivityGraphical}
\title{Overall productivity histogram}
\usage{
assessmentProductivityGraphical(
  ds = NULL,
  surveydate = NULL,
  dateformat = NULL,
  survey_consent = NULL
)
}
\arguments{
\item{ds}{dataset as a data.frame object}

\item{surveydate}{name as a string of the field in the dataset where the date of the survey is stored}

\item{dateformat}{format as a string used for the date ('\%m/\%d/\%Y')}

\item{survey_consent}{name as a string of the field in the dataset where the survey consent is stored}
}
\value{
graph  graphic as a plot.ly object
}
\description{
This function create an histogram showing the overall productivity per consent status per day.
}
\examples{
{
ds <- HighFrequencyChecks::sample_dataset
surveydate <- "survey_date"
dateformat <- "\%m/\%d/\%Y"
survey_consent <- "survey_consent"

assessmentProductivityGraphical(ds, surveydate, dateformat, survey_consent)
}
}
\author{
Yannick Pascaud
}
