% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_a.R
\name{assessmentTrackingSheet}
\alias{assessmentTrackingSheet}
\title{Overall tracking sheet}
\usage{
assessmentTrackingSheet(
  ds = NULL,
  dsSite = NULL,
  sampleSizeTable = NULL,
  sampleSizeTableSite = NULL,
  sampleSizeTableTarget = NULL,
  sampleSizeTableAvailable = NULL,
  surveyConsent = NULL,
  consentForValidSurvey = NULL
)
}
\arguments{
\item{ds}{dataset containing the survey (from kobo): data.frame}

\item{dsSite}{name of the field in the dataset where the site is stored: string}

\item{sampleSizeTable}{dataset containing the sampling frame: data.frame}

\item{sampleSizeTableSite}{name of the field in the sampling frame where the site is stored: string}

\item{sampleSizeTableTarget}{name of the field where the target number of survey is stored in the sampling frame: string}

\item{sampleSizeTableAvailable}{name of the field where the number of points generated is stored in the sampling frame: string}

\item{surveyConsent}{name of the field in the dataset where the survey consent is stored: string}

\item{consentForValidSurvey}{value defined in the kobo form to acknowledge the surveyed person gave his consent: string}
}
\value{
dst      same dataset as the inputed one but with survey marked for deletion if errors are found and delete=TRUE (or NULL)

ret_log  list of the errors found (or NULL)

var      a list of value (or NULL)

graph    graphical representation of the results (or NULL)
}
\description{
This function display the overall tracking sheet.
}
\examples{
{
ds <- HighFrequencyChecks::sample_dataset
dsSite <- "union_name"
sampleSizeTable <- HighFrequencyChecks::SampleSize
sampleSizeTableSite <- "Union"
sampleSizeTableTarget <- "SS"
sampleSizeTableAvailable <- "TotPts"                # Usually the Target + a buffer
surveyConsent <- "survey_consent"
consentForValidSurvey <- "yes"                      # consent value for yes

list[dst,ret_log,var,graph] <- assessmentTrackingSheet(ds=ds,
                                                 dsSite=dsSite,
                                                 sampleSizeTable=sampleSizeTable,
                                                 sampleSizeTableSite=sampleSizeTableSite,
                                                 sampleSizeTableTarget=sampleSizeTableTarget,
                                                 sampleSizeTableAvailable=sampleSizeTableAvailable,
                                                 surveyConsent=surveyConsent,
                                                 consentForValidSurvey=consentForValidSurvey)
head(ret_log,10)
print(graph)
}
}
\author{
Yannick Pascaud
}
