% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{booleanSum}
\alias{booleanSum}
\title{Sum bollean value}
\usage{
booleanSum(x)
}
\arguments{
\item{x}{vector of booleans}
}
\value{
boolean value
}
\description{
This function allow to perform a Boolean sum (AND) over a vector of boolean values (TRUE/FALSE)
}
