% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_e.R
\name{enumeratorErrorsDashboard}
\alias{enumeratorErrorsDashboard}
\title{Create a dashboard dislplaying the number of errors by enumerators}
\usage{
enumeratorErrorsDashboard(enumeratorID = NULL, reports = NULL)
}
\arguments{
\item{enumeratorID}{name of the field where the enumerator ID is stored: string}

\item{reports}{reports names generated from the other checks included in this package, be sure when you choose the columns to be
included in each report generated that the enumeratorID is selected before including the report as a parameter to this function:
list of string(c(report1,report2,...))}
}
\value{
dst      same dataset as the inputed one but with survey marked for deletion if errors are found and delete=TRUE (or NULL)

ret_log  list of the errors found (or NULL)

var      a list of value (or NULL)

graph    graphical representation of the results (or NULL)
}
\description{
This function display the number of errors made by the enumerator, one graph is generated by enumerator showing for each
}
\examples{
\dontrun{
enumeratorID <- "enumerator_id"
reports <- c("reportisInterviewAtTheSamplePoint",
             "reportisInterviewCompleted",
             "reportisInterviewInTheCorrectSite",
             "reportisInterviewTooShort",
             "reportisInterviewTooShortForTheHouseholdSize",
             "reportisInterviewWithConsent",
             "reportisSurveyEndBeforeItStarts",
             "reportisSurveyMadeInTheFuture",
             "reportisSurveyOnMoreThanADay",
             "reportisSurveyStartedBeforeTheAssessment",
             "reportisUniqueIDDuplicated",
             "reportisUniqueIDMissing")

list[dst,ret_log,var,graph] <- enumeratorErrorsDashboard(enumeratorID=enumeratorID,
                                                         reports=reports)
print(graph)
}
}
\author{
Yannick Pascaud
}
