% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{enumeratorProductivity}
\alias{enumeratorProductivity}
\title{Check the number of surveys by enumerator}
\usage{
enumeratorProductivity(ds = NULL, surveydate = NULL, enumeratorID = NULL)
}
\arguments{
\item{ds}{dataset as a data.frame object}

\item{surveydate}{name as a string of the field in the dataset where the date of the survey is stored}

\item{enumeratorID}{name as a string of the field in the dataset where the enumerator ID is stored}
}
\value{
ret_log  the report
}
\description{
This function display the total number of survey made and the average per day per enumerator.
}
\examples{
 {
ds <- HighFrequencyChecks::sample_dataset
surveydate <- "survey_date"
enumeratorID <- "enumerator_id"

list[dst,ret_log,var,graph] <- enumeratorProductivity(ds,
                                                      surveydate,
                                                      enumeratorID)
head(ret_log,10)
print(graph)
}
}
\author{
Yannick Pascaud
}
