% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{enumeratorSurveysDuration}
\alias{enumeratorSurveysDuration}
\title{Check the average interview duration by enumerator}
\usage{
enumeratorSurveysDuration(ds = NULL, dates = NULL, enumeratorID = NULL)
}
\arguments{
\item{ds}{dataset as a data.frame object}

\item{dates}{fields as a list of string where the survey start and end date is stored (c('start_date','end_date'))}

\item{enumeratorID}{name as a string of the field in the dataset where the enumerator ID is stored}
}
\value{
ret_log  the report
}
\description{
This function display the average interview duration per enumerator.
}
\examples{
{
ds <- HighFrequencyChecks::sample_dataset
dates <- c("survey_start","end_survey")
enumeratorID <- "enumerator_id"

list[dst,ret_log,var,graph] <- enumeratorSurveysDuration(ds,
                                                         dates,
                                                         enumeratorID)
head(ret_log,10)
print(graph)
}
}
\author{
Yannick Pascaud
}
