% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateRmd.R
\name{GenerateRmd}
\alias{GenerateRmd}
\title{Create an Rmd file for the report}
\usage{
GenerateRmd(
  usexlsform = FALSE,
  form = NULL,
  ds = NULL,
  surveyConsent = NULL,
  reportingColumns = NULL,
  uniqueID = NULL,
  dates = NULL,
  surveyDate = NULL,
  startDataCollection = NULL,
  dateFormat = NULL,
  minimumSurveyDuration = NULL,
  minimumSurveyDurationByIndividual = NULL,
  sdval = NULL,
  HouseholdSize = NULL,
  deleteIsInterviewCompleted = NULL,
  deleteIsInterviewWithConsent = NULL,
  correctIsInterviewInTheCorrectSite = NULL,
  deleteIsInterviewAtTheSamplePoint = NULL,
  deleteIsUniqueIDMissing = NULL,
  deleteIsUniqueIDDuplicated = NULL,
  deleteIsSurveyOnMoreThanADay = NULL,
  deleteIsSurveyEndBeforeItStarts = NULL,
  deleteIsSurveyStartedBeforeTheAssessment = NULL,
  deleteIsSurveyMadeInTheFuture = NULL,
  deleteIsInterviewTooShort = NULL,
  deleteIsInterviewTooShortForTheHouseholdSize = NULL,
  enumeratorID = NULL,
  enumeratorcheck = FALSE,
  otherPattern = NULL,
  adminBoundaries = NULL,
  dsSite = NULL,
  dsCoordinates = NULL,
  adminBoundariesSite = NULL,
  sampledPoints = NULL,
  buffer = NULL,
  questionsSurveyBigValues = NULL,
  questionsSurveySmallValues = NULL,
  questionsEnumeratorIsLazy = NULL,
  sampleSizeTable = NULL,
  sampleSizeTableSite = NULL,
  sampleSizeTableTarget = NULL,
  sampleSizeTableAvailable = NULL,
  consentForValidSurvey = NULL
)
}
\arguments{
\item{usexlsform}{do not know but have to be documented}

\item{form}{do not know but have to be documented}

\item{ds}{dataset as a data.frame object}

\item{surveyConsent}{name as a string of the field in the dataset where the survey consent is stored}

\item{reportingColumns}{columns as a list of string name from the dataset you want in the result (c('col1','col2',...))}

\item{uniqueID}{name as a string of the field in the dataset where the unique ID is stored}

\item{dates}{fields as a list of string where the survey start and end date is stored (c('start_date','end_date'))}

\item{surveyDate}{name as a string of the field in the dataset where the date of the survey is stored}

\item{startDataCollection}{date as a string of the first day of data collection ('yyyy-mm-dd')}

\item{dateFormat}{format as a string used for the date ('\%m/\%d/\%Y')}

\item{minimumSurveyDuration}{minimum acceptable survey duration as integer in minutes}

\item{minimumSurveyDurationByIndividual}{minimum acceptable survey duration for one indifidual as integer in minutes}

\item{sdval}{number of standard deviation for which the data within is considered as acceptable}

\item{HouseholdSize}{name as a string of the field in the dataset where the household size is stored}

\item{deleteIsInterviewCompleted}{delete action to be done as a boolean (TRUE/FALSE) for isInterviewCompleted function}

\item{deleteIsInterviewWithConsent}{delete action to be done as a boolean (TRUE/FALSE) for isInterviewWithConsent function}

\item{correctIsInterviewInTheCorrectSite}{correction action to be done as a boolean (TRUE/FALSE) for isInterviewInTheCorrectSite function}

\item{deleteIsInterviewAtTheSamplePoint}{delete action to be done as a boolean (TRUE/FALSE) for isInterviewAtTheSamplePoint function}

\item{deleteIsUniqueIDMissing}{delete action to be done as a boolean (TRUE/FALSE) for isUniqueIDMissing function}

\item{deleteIsUniqueIDDuplicated}{delete action to be done as a boolean (TRUE/FALSE) for isUniqueIDDuplicated function}

\item{deleteIsSurveyOnMoreThanADay}{delete action to be done as a boolean (TRUE/FALSE) for isSurveyOnMoreThanADay function}

\item{deleteIsSurveyEndBeforeItStarts}{delete action to be done as a boolean (TRUE/FALSE) for isSurveyEndBeforeItStarts function}

\item{deleteIsSurveyStartedBeforeTheAssessment}{delete action to be done as a boolean (TRUE/FALSE) for isSurveyStartedBeforeTheAssessment function}

\item{deleteIsSurveyMadeInTheFuture}{delete action to be done as a boolean (TRUE/FALSE) for isSurveyMadeInTheFuture function}

\item{deleteIsInterviewTooShort}{delete action to be done as a boolean (TRUE/FALSE) for isInterviewTooShort function}

\item{deleteIsInterviewTooShortForTheHouseholdSize}{delete action to be done as a boolean (TRUE/FALSE) for isInterviewTooShortForTheHouseholdSize function}

\item{enumeratorID}{name as a string of the field in the dataset where the enumerator ID is stored}

\item{enumeratorcheck}{specify if the report has to be displayed for each enumerator or not as a boolean (TRUE/FALSE)}

\item{otherPattern}{pattern as string to identify the fields containing others values ('_other$')}

\item{adminBoundaries}{dataset containing the shapefile - Regardless the projection used for the shapefile, it is transformed to WGS84}

\item{dsSite}{name as a string of the field in the dataset where the site is stored}

\item{dsCoordinates}{columns as a list of string name from the dataset where the GPS coordinates are stored (c('Long','Lat'))}

\item{adminBoundariesSite}{name as a string of the field in the shapefile where the site is stored}

\item{sampledPoints}{dataset containing the shapefile - Regardless the projection used for the shapefile, it is transformed to WGS84}

\item{buffer}{value as an integer in meter to determine the buffer from a sampled point which is acceptable}

\item{questionsSurveyBigValues}{columns as a list of string name from the dataset you want to check against (c(col1=x,col2=x,...))}

\item{questionsSurveySmallValues}{columns as a list of string name from the dataset you want to check against (c(col1=x,col2=x,...))}

\item{questionsEnumeratorIsLazy}{columns as a list of string name from the dataset you want to check against (c(col1=x,col2=x,...))}

\item{sampleSizeTable}{sampling frame as a data.frame object}

\item{sampleSizeTableSite}{name as a string of the field in the sampling frame where the site is stored}

\item{sampleSizeTableTarget}{name as a string of the field where the target number of survey is stored in the sampling frame}

\item{sampleSizeTableAvailable}{name as a string of the field where the number of points generated is stored in the sampling frame}

\item{consentForValidSurvey}{consent value for the valid surveys as string}
}
\value{
vignette file
}
\description{
Create an Rmd file for the report
}
\examples{
\dontrun{
ds <- HighFrequencyChecks::sample_dataset

survey_consent <- "survey_consent"
reportingcol <- c("enumerator_id","X_uuid")
UniqueID <- "X_uuid"

## Setting dates
dates <- c("survey_start","end_survey")
surveydate <- "survey_date"
start_collection <- "2018-11-11"
dateformat <- "\%m/\%d/\%Y"
minduration <- 30
sdval <- 2
HHSize <-"consent_received.respondent_info.hh_size"


## Setting action on dataset
delete <- FALSE
correct <- FALSE

## Setting enumerator
enumeratorID <- "enumerator_id"
enumeratorcheck <- FALSE
otherpattern <- "_other$"

## Setting linked geodata
adm <- HighFrequencyChecks::admin
ds_site <- "union_name"
ds_coord <- c("X_gps_reading_longitude","X_gps_reading_latitude")
adm_site <- "Union"
pts <- HighFrequencyChecks::SamplePts
buff <- 10

## setting questions to check
questions <-c("consent_received.food_security.spend_food",
  "consent_received.food_security.spend_medication",
  "consent_received.food_security.spend_education",
  "consent_received.food_security.spend_fix_shelter",
  "consent_received.food_security.spend_clothing",
  "consent_received.food_security.spend_hygiene",
  "consent_received.food_security.spend_fuel",
  "consent_received.food_security.spend_hh_items",
  "consent_received.food_security.spend_transport",
  "consent_received.food_security.spend_communication",
  "consent_received.food_security.spend_tobacco",
  "consent_received.food_security.spend_rent",
  "consent_received.food_security.spend_debts",
  "consent_received.food_security.spend_other")
value <- 25000
minnbanswers <- 3


## setting sampling plan
sf <- HighFrequencyChecks::SampleSize
dssite <- "union_name"
sfsite <- "Union"
survey_consent <- "survey_consent"
sftarget <- "SS"
sfnbpts <- "TotPts"
formul <- c("done-no-not_eligible-delete",
"done-no-not_eligible-deleted-SS")
colorder <- c("site","SS","Provisio","done","not_eligible",
 "no","deleted","yes","final","variance")

GenerateRmd(ds,
  survey_consent,
  reportingcol,
  UniqueID,
  dates,
  surveydate,
  start_collection,
  dateformat,
  minduration,
  sdval,
  HHSize,
  delete,
  correct,
  enumeratorID,
  enumeratorcheck,
  otherpattern,
  adm,
  ds_site,
  ds_coord,
  adm_site,
  pts,
  buff,
  questions,
  value,
  minnbanswers,
  sf, dssite,
  sfsite,
  sftarget,
  sfnbpts,
  formul,
  colorder)
}
}
\author{
Yannick Pascaud, Edouard Legoupil
}
