% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfc_projectinit.R
\name{hfc_projectinit}
\alias{hfc_projectinit}
\title{Analysis package project initiation}
\usage{
hfc_projectinit()
}
\value{
A structure of directory and scripts in order to set up quickly a project.
}
\description{
Create analysis package project structure
}
\examples{
\dontrun{
hfc_projectinit()
}
}
\author{
Edouard Legoupil, Elliott Messeiller
}
