% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfcApp.R
\name{hfcApp}
\alias{hfcApp}
\title{starts the graphical user interface developed with \emph{shiny}.}
\usage{
hfcApp(
  maxRequestSize = 50,
  debug = FALSE,
  theme = "yeti",
  ...,
  shiny.server = FALSE
)
}
\arguments{
\item{maxRequestSize}{(numeric) number defining the maximum allowed filesize (in megabytes)
for uploaded files, defaults to 50MB}

\item{debug}{logical if \code{TRUE}, set shiny-debugging options}

\item{theme}{select stylesheet for the interface. Supported choices are
\itemize{
\item 'yeti'
\item 'flatly'
}}

\item{...}{arguments (e.g \code{host}) that are passed through \code{\link[shiny]{runApp}} when
starting the shiny application}

\item{shiny.server}{Setting this parameter to \code{TRUE} will return the app in the form of an
object rather than invoking it. This is useful for deploying \code{hfcApp} via \code{shiny-server}.}
}
\value{
starts the interactive graphical user interface which may be used to perform the
anonymisation process.
}
\description{
starts the graphical user interface developed with \emph{shiny}.
}
\examples{
\dontrun{
hfcApp(theme="flatly")
}
}
