% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfcProjectInit.R
\name{hfcProjectInit}
\alias{hfcProjectInit}
\title{Analysis package project initiation}
\usage{
hfcProjectInit()
}
\value{
A structure of directory and scripts in order to set up quickly a project.
}
\description{
Create analysis package project structure
}
\examples{
\dontrun{
hfcProjectInit()
}
}
\author{
Edouard Legoupil, Elliott Messeiller
}
