% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_i.R
\name{isInterviewCompleted}
\alias{isInterviewCompleted}
\title{Check that all interviews were completed}
\usage{
isInterviewCompleted(
  ds = NULL,
  surveyConsent = NULL,
  dates = NULL,
  reportingColumns = c(enumeratorID, uniqueID),
  deleteIsInterviewCompleted = FALSE
)
}
\arguments{
\item{ds}{dataset containing the survey (from kobo): data.frame}

\item{surveyConsent}{name of the field in the dataset where the survey consent is stored: string}

\item{dates}{name of the fields where the information about the start and end date of the survey is stored: list of string (c('start_date','end_date'))}

\item{reportingColumns}{(Optional, by default it is built from the enumeratorID and the UniqueID) name of the columns from the dataset you want in the result: list of string (c('col1','col2',...))}

\item{deleteIsInterviewCompleted}{(Optional, by default set as FALSE) if TRUE, the survey in error will be marked as 'deletedIsInterviewCompleted': boolean (TRUE/FALSE)}

\item{uniqueID}{name of the field where the survey unique ID is stored: string}

\item{enumeratorID}{name of the field where the enumerator ID is stored: string}
}
\value{
dst      same dataset as the inputed one but with survey marked for deletion if errors are found and delete=TRUE (or NULL)

ret_log  list of the errors found (or NULL)

var      a list of value (or NULL)

graph    graphical representation of the results (or NULL)
}
\description{
This function check that all interviews in the dataset are completed,
  meaning all the interviews have an end date and time.
  There is an option to automatically mark for deletion the surveys which have not an end date.
}
\examples{
{
ds <- HighFrequencyChecks::sample_dataset
surveyConsent <- "survey_consent"
dates <- c("survey_start","end_survey")
uniqueID <- "X_uuid"
enumeratorID <- "enumerator_id"
reportingColumns <- c(enumeratorID, uniqueID)

list[dst,ret_log,var,graph] <- isInterviewCompleted(ds=ds,
                                                    surveyConsent=surveyConsent,
                                                    dates=dates,
                                                    reportingColumns=reportingColumns,
                                                    deleteIsInterviewCompleted=FALSE)
head(ret_log,10)
print(graph)
}

}
\author{
Yannick Pascaud
}
