% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_i.R
\name{isInterviewTooShortForTheHouseholdSize}
\alias{isInterviewTooShortForTheHouseholdSize}
\title{Check that the duration relative to the household size of each interview is more than a threshold}
\usage{
isInterviewTooShortForTheHouseholdSize(
  ds = NULL,
  surveyConsent = NULL,
  dates = NULL,
  householdSize = NULL,
  minimumSurveyDurationByIndividual = 10,
  reportingColumns = c(enumeratorID, uniqueID),
  deleteIsInterviewTooShortForTheHouseholdSize = FALSE
)
}
\arguments{
\item{ds}{dataset containing the survey (from kobo): data.frame}

\item{surveyConsent}{name of the field in the dataset where the survey consent is stored: string}

\item{dates}{name of the fields where the information about the start and end date of the survey is stored: list of string (c('start_date','end_date'))}

\item{householdSize}{name of the field in the dataset where the household size is stored: string}

\item{minimumSurveyDurationByIndividual}{minimum acceptable survey duration for one individual in minutes: integer}

\item{reportingColumns}{(Optional, by default it is built from the enumeratorID and the UniqueID) name of the columns from the dataset you want in the result: list of string (c('col1','col2',...))}

\item{deleteIsInterviewTooShortForTheHouseholdSize}{(Optional, by default set as FALSE) if TRUE, the survey in error will be marked as 'deletedIsInterviewTooShortForTheHouseholdSize': boolean (TRUE/FALSE)}

\item{uniqueID}{name of the field where the survey unique ID is stored: string}

\item{enumeratorID}{name of the field where the enumerator ID is stored: string}
}
\value{
dst      same dataset as the inputed one but with survey marked for deletion if errors are found and delete=TRUE (or NULL)

ret_log  list of the errors found (or NULL)

var      a list of value (or NULL)

graph    graphical representation of the results (or NULL)
}
\description{
This function check that the duration relative to the household size of each interview is more than a specified threshold.
There is an option to automatically mark for deletion the surveys which are under the threshold.
Warning: If there are uncorrected mistakes in the survey dates, it can lead to have the length of the survey in seconds and this check will not performed well
}
\examples{
{
ds <- HighFrequencyChecks::sample_dataset
surveyConsent <- "survey_consent"
dates <- c("survey_start","end_survey")
householdSize <-"consent_received.respondent_info.hh_size"
uniqueID <- "X_uuid"
enumeratorID <- "enumerator_id"
minimumSurveyDurationByIndividual <- 10
reportingColumns <- c(enumeratorID, uniqueID)

list[dst,ret_log,var,graph] <- isInterviewTooShortForTheHouseholdSize(ds=ds,
                           surveyConsent=surveyConsent,
                           dates=dates,
                           householdSize=householdSize,
                           minimumSurveyDurationByIndividual=minimumSurveyDurationByIndividual,
                           reportingColumns=reportingColumns,
                           deleteIsInterviewTooShortForTheHouseholdSize=FALSE)
head(ret_log, 10)
print(graph)
}
}
\author{
Yannick Pascaud
}
