% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{isSurveyMadeInTheFuture}
\alias{isSurveyMadeInTheFuture}
\title{Surveys that have start date/time after system date}
\usage{
isSurveyMadeInTheFuture(
  ds = NULL,
  survey_consent = NULL,
  dates = NULL,
  reportingcol = NULL,
  delete = NULL
)
}
\arguments{
\item{ds}{dataset as a data.frame object}

\item{survey_consent}{name as a string of the field in the dataset where the survey consent is stored}

\item{dates}{fields as a list of string where the survey start and end date is stored (c('start_date','end_date'))}

\item{reportingcol}{columns as a list of string name from the dataset you want in the result (c('col1','col2',...))}

\item{delete}{delete action to be done as a boolean (TRUE/FALSE)}
}
\value{
dst same dataset as the inputed one but with survey marked for deletion if errors are found and delete=TRUE

ret_log  list of the errors found
}
\description{
This function check that all interviews in the dataset do not start after the current date.
There is an option to automatically mark for deletion the surveys which have a start date in the future.
}
\examples{
{
ds <- HighFrequencyChecks::sample_dataset
dates <- c("survey_start","end_survey")
survey_consent <- "survey_consent"
reportingcol <- c("enumerator_id","X_uuid")
delete <- FALSE


list[dst,ret_log,var,graph] <- isSurveyMadeInTheFuture(ds,
                                              survey_consent,
                                              dates,
                                              reportingcol,
                                              delete)
head(ret_log, 10)
}
}
\author{
Yannick Pascaud
}
