% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_i.R
\name{isSurveyStartedBeforeTheAssessment}
\alias{isSurveyStartedBeforeTheAssessment}
\title{Surveys that show start date earlier than first day of data collection}
\usage{
isSurveyStartedBeforeTheAssessment(
  ds = NULL,
  dates = NULL,
  surveyConsent = NULL,
  startDataCollection = NULL,
  reportingColumns = c(enumeratorID, uniqueID),
  deleteIsSurveyStartedBeforeTheAssessment = FALSE
)
}
\arguments{
\item{ds}{dataset containing the survey (from kobo): data.frame}

\item{dates}{name of the fields where the information about the start and end date of the survey is stored: list of string (c('start_date','end_date'))}

\item{surveyConsent}{name of the field in the dataset where the survey consent is stored: string}

\item{startDataCollection}{date of the first day of the data collection: string ('yyyy-mm-dd')}

\item{reportingColumns}{(Optional, by default it is built from the enumeratorID and the UniqueID) name of the columns from the dataset you want in the result: list of string (c('col1','col2',...))}

\item{deleteIsSurveyStartedBeforeTheAssessment}{(Optional, by default set as FALSE) if TRUE, the survey in error will be marked as 'deletedIsSurveyStartedBeforeTheAssessment': boolean (TRUE/FALSE)}

\item{uniqueID}{name of the field where the survey unique ID is stored: string}

\item{enumeratorID}{name of the field where the enumerator ID is stored: string}
}
\value{
dst      same dataset as the inputed one but with survey marked for deletion if errors are found and delete=TRUE (or NULL)

ret_log  list of the errors found (or NULL)

var      a list of value (or NULL)

graph    graphical representation of the results (or NULL)
}
\description{
This function check that all interviews in the dataset start after the actual first day of data collection.
There is an option to automatically mark for deletion the surveys which have started before the first day of data collection.
}
\examples{
{
ds <- HighFrequencyChecks::sample_dataset
dates <- c("survey_start","end_survey")
surveyConsent <- "survey_consent"
startDataCollection <- "2018-11-11"
uniqueID <- "X_uuid"
enumeratorID <- "enumerator_id"
reportingColumns <- c(enumeratorID, uniqueID)

list[dst,ret_log,var,graph] <- isSurveyStartedBeforeTheAssessment(ds=ds,
                                                   dates=dates,
                                                   surveyConsent=surveyConsent,
                                                   startDataCollection=startDataCollection,
                                                   reportingColumns=reportingColumns,
                                                   deleteIsSurveyStartedBeforeTheAssessment = FALSE)
head(ret_log, 10)
}
}
\author{
Yannick Pascaud
}
