% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mapFunctions}
\alias{mapFunctions}
\title{Mapping of the function which can be used for the report}
\usage{
mapFunctions(variablesConfig, reportType)
}
\arguments{
\item{variablesConfig}{dataset where the variables are defined}

\item{reportType}{type of report wanted (N for text one, G for graphical one)}
}
\value{
list of possible functions
}
\description{
Mapping of the function which can be used for the report based on the variables configured
}
