% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ni}
\alias{ni}
\alias{\%ni\%}
\title{Not in}
\usage{
x \%ni\% table
}
\arguments{
\item{x}{vector or NULL: the values to be matched. Long vectors are supported.}

\item{table}{vector or NULL: the values to be matched against. Long vectors are not supported.}
}
\description{
This function create the not in function
}
