% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{piechart}
\alias{piechart}
\title{Create a donut chart with ggplot}
\usage{
piechart(data, graphTitle)
}
\arguments{
\item{data}{dataset as a data.frame object}

\item{graphTitle}{aes paramters}
}
\value{
a ggplot object
}
\description{
This function allow to create a donut chart with ggplot.
}
