% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportWrapper.R
\name{ReportWrapper}
\alias{ReportWrapper}
\title{Create an Rmd file for the report}
\usage{
ReportWrapper(
  working_directY = NULL,
  xlsform_name = NULL,
  xlsform_location = NULL,
  vignette_directory = NULL
)
}
\arguments{
\item{working_directY}{the current working directory}

\item{xlsform_name}{the name of the xlsform used to build the report}

\item{xlsform_location}{the location of the xlsform from the working directory}

\item{vignette_directory}{the directory where to generate the .Rmd from the working directory}
}
\value{
vignette file
}
\description{
Create an Rmd file for the report
}
\examples{
\dontrun{
working_directY <- getwd()
xlsform_name <- "form.xlsx"
xlsform_location <- "/inst/demo/"
vignette_directory <- "/vignettes/HFC_WR.Rmd"

ReportWrapper(working_directY,
              xlsform_name,
              xlsform_location,
              vignette_directory)
}
}
\author{
Yannick Pascaud
}
