\name{sample_dataset}
\alias{sample_dataset}
\docType{data}
\title{
household dataset to be used as an example
}
\description{
This dataset comes from the Host Community Multi Sector Needs Assessment conducted in 2018 by IMPACT/REACH on behalf of the ISCG and funded by ECHO. This is a sample of the data collected in both Ukhia and Teknaf.

For privacy purpose, the GPS coordinates are faked but are still located in the designated areas.
}
\usage{sample_dataset}
\format{
  A data frame with 498 observations on 587 variables.
}
\keyword{datasets}
