\name{SamplePts}
\alias{SamplePts}
\docType{data}
\title{
points sampled to be surveyed dataset to be used as an example
}
\description{
This dataset is a sample of the points to be surveyed in both Ukhia and Teknaf.

For privacy purpose, the GPS coordinates are faked but are still located in the designated areas.
}
\usage{sample_dataset}
\format{
  spatial dataset
}
\keyword{datasets}
