\name{SampleSize}
\alias{SampleSize}
\docType{data}
\title{
sample size used
}
\description{
This dataset comes from the Host Community Multi Sector Needs Assessment conducted in 2018 by IMPACT/REACH on behalf of the ISCG and funded by ECHO. This is the sample size per union used in both Ukhia and Teknaf.
}
\usage{sample_dataset}
\format{
  A data frame with 11 observations on 7 variables.
}
\keyword{datasets}
