% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_s.R
\name{surveyBigValues}
\alias{surveyBigValues}
\title{Report the values greater than a specified value per specified fields}
\usage{
surveyBigValues(
  ds = NULL,
  questionsSurveyBigValues = NULL,
  enumeratorID = NULL,
  reportingColumns = c(enumeratorID, uniqueID),
  enumeratorCheck = FALSE
)
}
\arguments{
\item{ds}{dataset containing the survey (from kobo): data.frame}

\item{questionsSurveyBigValues}{columns name from the dataset and value you want to check against (c(col1=value1,col2=value2,...)): named list of integer}

\item{enumeratorID}{name of the field where the enumerator ID is stored: string}

\item{reportingColumns}{(Optional, by default it is built from the enumeratorID and the UniqueID) name of the columns from the dataset you want in the result: list of string (c('col1','col2',...))}

\item{enumeratorCheck}{(Optional, by default set to FALSE) specify if the report has to be displayed for each enumerator or not: boolean (TRUE/FALSE)}

\item{uniqueID}{name of the field where the survey unique ID is stored: string}
}
\value{
dst      same dataset as the inputed one but with survey marked for deletion if errors are found and delete=TRUE (or NULL)

ret_log  list of the errors found (or NULL)

var      a list of value (or NULL)

graph    graphical representation of the results (or NULL)
}
\description{
This function provide a report showing all values which are greater than a certain threshold for a specified list of fields.
}
\examples{
{
ds <- HighFrequencyChecks::sample_dataset
questionsSurveyBigValues <-c(consent_received.food_security.spend_food=25000,
                             consent_received.food_security.spend_medication=25000,
                             consent_received.food_security.spend_education=25000,
                             consent_received.food_security.spend_fix_shelter=25000,
                             consent_received.food_security.spend_clothing=25000,
                             consent_received.food_security.spend_hygiene=25000,
                             consent_received.food_security.spend_fuel=25000,
                             consent_received.food_security.spend_hh_items=25000,
                             consent_received.food_security.spend_transport=25000,
                             consent_received.food_security.spend_communication=25000,
                             consent_received.food_security.spend_tobacco=25000,
                             consent_received.food_security.spend_rent=25000,
                             consent_received.food_security.spend_debts=25000,
                             consent_received.food_security.spend_other=25000)
enumeratorID <- "enumerator_id"
uniqueID <- "X_uuid"
reportingColumns <- c(enumeratorID, uniqueID)

list[dst,ret_log,var,graph] <- surveyBigValues(ds=ds,
                                               questionsSurveyBigValues=questionsSurveyBigValues,
                                               enumeratorID=enumeratorID,
                                               reportingColumns=reportingColumns,
                                               enumeratorCheck=FALSE)
head(ret_log,10)
}
}
\author{
Yannick Pascaud
}
