% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_s.R
\name{surveyDistinctValues}
\alias{surveyDistinctValues}
\title{Number of distinct values (not missing) per fields}
\usage{
surveyDistinctValues(ds = NULL, enumeratorID = NULL, enumeratorCheck = FALSE)
}
\arguments{
\item{ds}{dataset containing the survey (from kobo): data.frame}

\item{enumeratorID}{name of the field where the enumerator ID is stored: string}

\item{enumeratorCheck}{(Optional, by default set to FALSE) specify if the report has to be displayed for each enumerator or not: boolean (TRUE/FALSE)}
}
\value{
dst      same dataset as the inputed one but with survey marked for deletion if errors are found and delete=TRUE (or NULL)

ret_log  list of the errors found (or NULL)

var      a list of value (or NULL)

graph    graphical representation of the results (or NULL)
}
\description{
This function provide a report showing the number of distinct values for each fields.
This report can be global (all the surveys) or displayed for each enumerator ID
}
\examples{
{
ds <- HighFrequencyChecks::sample_dataset
enumeratorID <- "enumerator_id"

list[dst,ret_log,var,graph] <- surveyDistinctValues(ds=ds,
                                                    enumeratorID=enumeratorID,
                                                    enumeratorCheck=FALSE)
head(ret_log,10)
}
}
\author{
Yannick Pascaud
}
