% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_s.R
\name{surveyOtherValues}
\alias{surveyOtherValues}
\title{List of other distinct values (not missing) per fields other with count}
\usage{
surveyOtherValues(
  ds = NULL,
  otherPattern = NULL,
  enumeratorID = NULL,
  enumeratorCheck = FALSE
)
}
\arguments{
\item{ds}{dataset containing the survey (from kobo): data.frame}

\item{otherPattern}{patternto identify the fields containing others values (eg: '_other$'): string}

\item{enumeratorID}{name of the field where the enumerator ID is stored: string}

\item{enumeratorCheck}{(Optional, by default set to FALSE) specify if the report has to be displayed for each enumerator or not: boolean (TRUE/FALSE)}
}
\value{
dst      same dataset as the inputed one but with survey marked for deletion if errors are found and delete=TRUE (or NULL)

ret_log  list of the errors found (or NULL)

var      a list of value (or NULL)

graph    graphical representation of the results (or NULL)
}
\description{
This function provide a report showing all distinct other values and the number of occurrences for each fields "other".
This report can be global (all the surveys) or displayed for each enumerator ID
}
\examples{
{
ds <- HighFrequencyChecks::sample_dataset
otherPattern <- "_other$"
enumeratorID <- "enumerator_id"

list[dst,ret_log,var,graph] <- surveyOtherValues(ds=ds,
                                                 otherPattern=otherPattern,
                                                 enumeratorID=enumeratorID,
                                                 enumeratorCheck=FALSE)
head(ret_log,10)
}
}
\author{
Yannick Pascaud
}
