% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{assessmentDailyValidSurveys}
\alias{assessmentDailyValidSurveys}
\title{Daily number of survey per consent status}
\usage{
assessmentDailyValidSurveys(
  ds = NULL,
  surveydate = NULL,
  dateformat = NULL,
  survey_consent = NULL
)
}
\arguments{
\item{ds}{dataset as a data.frame object}

\item{surveydate}{name as a string of the field in the dataset where the date of the survey is stored}

\item{dateformat}{format as a string used for the date ('\%m/\%d/\%Y')}

\item{survey_consent}{name as a string of the field in the dataset where the survey consent is stored}
}
\value{
ret_log  the report
}
\description{
This function display the number of surveys conducted per day per constent status.
}
\examples{
{
ds <- HighFrequencyChecks::sample_dataset
surveydate <- "survey_date"
dateformat <- "\%m/\%d/\%Y"
survey_consent <- "survey_consent"

list[dst,ret_log,var,graph] <- assessmentDailyValidSurveys(ds,
                                                           surveydate,
                                                           dateformat,
                                                           survey_consent)
head(ret_log,10)
print(graph)
}

}
\author{
Yannick Pascaud
}
