% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{assessmentDurationOutliers}
\alias{assessmentDurationOutliers}
\title{Report the outlier durations for the surveys}
\usage{
assessmentDurationOutliers(
  ds = NULL,
  dates = NULL,
  sdval = NULL,
  reportingcol = NULL
)
}
\arguments{
\item{ds}{dataset as a data.frame object}

\item{dates}{fields as a list of string where the survey start and end date is stored (c('start_date','end_date'))}

\item{sdval}{number of standard deviation for which the duration within is considered as acceptable}

\item{reportingcol}{columns as a list of string name from the dataset you want in the result (c('col1','col2',...))}
}
\value{
logf  the report
}
\description{
This function report the outlier durations for the surveys
}
\examples{
{
ds <- HighFrequencyChecks::sample_dataset
dates <- c("survey_start","end_survey")
sdval <- 5
reportingcol <- c("enumerator_id","X_uuid")

log <- assessmentDurationOutliers(ds, dates, sdval, reportingcol)
head(log,10)
}
}
\author{
Yannick Pascaud
}
