% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{assessmentTrackingSheet}
\alias{assessmentTrackingSheet}
\title{Overall tracking sheet}
\usage{
assessmentTrackingSheet(
  ds = NULL,
  sf = NULL,
  dssite = NULL,
  sfsite = NULL,
  survey_consent = NULL,
  sftarget = NULL,
  sfnbpts = NULL,
  formul = NULL,
  colorder = NULL
)
}
\arguments{
\item{ds}{dataset as a data.frame object}

\item{sf}{sampling frame as a data.frame object}

\item{dssite}{name as a string of the field in the dataset where the site is stored}

\item{sfsite}{name as a string of the field in the sampling frame where the site is stored}

\item{survey_consent}{name as a string of the field in the dataset where the survey consent is stored}

\item{sftarget}{name as a string of the field where the target number of survey is stored in the sampling frame}

\item{sfnbpts}{name as a string of the field where the number of points generated is stored in the sampling frame}

\item{formul}{formulas as a list of string used to compute the final number of eligible surveys and the variance from the target (C('formula1','formula2')).
the values/fields available are: done and the ones generated according the survey consent values (one per value)}

\item{colorder}{column names as a list of string to order the colums in the result (C('col1','col2',...)).
the columns available are: site, done, final, variance and the ones generated according the survey consent values (one per value)}
}
\value{
logf  the report
}
\description{
This function display the overall tracking sheet.
}
\examples{
\dontrun{
ds <- HighFrequencyChecks::sample_dataset
sf <- HighFrequencyChecks::SampleSize
dssite <- "union_name"
sfsite <- "Union"
survey_consent <- "survey_consent"
sftarget <- "SS"
sfnbpts <- "TotPts"
formul <- c("done-no-not_eligible-delete",
               "done-no-not_eligible-deleted-SS")
colorder <- c("site","SS","Provisio","done","not_eligible",
                 "no","deleted","yes","final","variance")

log <- assessmentTrackingSheet(ds,
                        sf,
                        dssite,
                        sfsite,
                        survey_consent,
                        sftarget,
                        sfnbpts,
                        formul,
                        colorder)
head(log,10)

}
}
\author{
Yannick Pascaud
}
