% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFrequencyChecksSection1.R
\name{chk1a_interview_completed}
\alias{chk1a_interview_completed}
\title{Check that all interviews were completed}
\usage{
chk1a_interview_completed(
  ds = NULL,
  survey_consent = NULL,
  dates = NULL,
  reportingcol = NULL,
  delete = NULL
)
}
\arguments{
\item{ds}{dataset as a data.frame object}

\item{survey_consent}{name as a string of the field in the dataset where the survey consent is stored}

\item{dates}{fields as a list of string where the survey start and end date is stored (c('start_date','end_date'))}

\item{reportingcol}{columns as a list of string name from the dataset you want in the result (c('col1','col2',...))}

\item{delete}{delete action to be done as a boolean (TRUE/FALSE)}
}
\value{
ds same dataset as the inputed one but with survey marked for deletion if errors are found and delete=TRUE

errors  list of the errors found
}
\description{
This function check that all interviews in the dataset are completed,
  meaning all the interviews have an end date and time.
  There is an option to automatically mark for deletion the surveys which have not an end date.
}
\examples{
{
ds <- HighFrequencyChecks::sample_dataset
survey_consent <- "survey_consent"
dates <- c("survey_start","end_survey")
reportingcol <- c("enumerator_id","X_uuid")
delete <- FALSE

head(chk1a_interview_completed(ds, survey_consent, dt, reportingcol, delete),10)
}

}
\author{
Yannick Pascaud
}
