% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFrequencyChecksSection1.R
\name{chk1b_survey_consent}
\alias{chk1b_survey_consent}
\title{Check that all surveys have consent}
\usage{
chk1b_survey_consent(
  ds = NULL,
  survey_consent = NULL,
  reportingcol = NULL,
  delete = NULL
)
}
\arguments{
\item{ds}{dataset as a data.frame object}

\item{survey_consent}{name as a string of the field in the dataset where the survey consent is stored}

\item{reportingcol}{columns as a list of string name from the dataset you want in the result (c('col1','col2',...))}

\item{delete}{delete action to be done as a boolean (TRUE/FALSE)}
}
\value{
ds same dataset as the inputed one but with survey marked for deletion if errors are found and delete=TRUE

errors  list of the errors found
}
\description{
This function check that all interviews in the dataset have information about the consent
 of the people surveyed, meaning all the field where this information is stored is not empty.
 There is an option to automatically mark for deletion the surveys which have not consent information.
}
\examples{
{
df <- HighFrequencyChecks::sample_dataset
sc <- "survey_consent"
reportingcol <- c("enumerator_id","X_uuid")
delete <- TRUE

head(chk1b_survey_consent(df, sc, dt, reportingcol, delete),10)
}
}
\author{
Yannick Pascaud
}
