% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFrequencyChecksSection3.R
\name{chk3a_date_mistake}
\alias{chk3a_date_mistake}
\title{Surveys that do not end on the same day as they started}
\usage{
chk3a_date_mistake(
  ds = NULL,
  survey_consent = NULL,
  dates = NULL,
  reportingcol = NULL,
  delete = NULL
)
}
\arguments{
\item{ds}{dataset as a data.frame object}

\item{survey_consent}{name as a string of the field in the dataset where the survey consent is stored}

\item{dates}{fields as a list of string where the survey start and end date is stored (c('start_date','end_date'))}

\item{reportingcol}{columns as a list of string name from the dataset you want in the result (c('col1','col2',...))}

\item{delete}{delete action to be done as a boolean (TRUE/FALSE)}
}
\value{
ds same dataset as the inputed one but with survey marked for deletion if errors are found and delete=TRUE

errors  list of the errors found
}
\description{
This function check that all interviews in the dataset start and end the same day.
There is an option to automatically mark for deletion the surveys which have different starting and ending dates.
}
\examples{
{
df <- HighFrequencyChecks::sample_dataset
sc <- "survey_consent"
dates <- c("survey_start","end_survey")
rc <- c("enumerator_id","X_uuid")
dl <- FALSE


head(chk3a_date_mistake(df, sc, dates, rc, dl),10)
}
}
\author{
Yannick Pascaud
}
