% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFrequencyChecksSection3.R
\name{chk3c_date_mistake}
\alias{chk3c_date_mistake}
\title{Surveys that show start date earlier than first day of data collection}
\usage{
chk3c_date_mistake(
  ds = NULL,
  survey_consent = NULL,
  dates = NULL,
  start_collection = NULL,
  reportingcol = NULL,
  delete = NULL
)
}
\arguments{
\item{ds}{dataset as a data.frame object}

\item{survey_consent}{name as a string of the field in the dataset where the survey consent is stored}

\item{dates}{fields as a list of string where the survey start and end date is stored (c('start_date','end_date'))}

\item{reportingcol}{columns as a list of string name from the dataset you want in the result (c('col1','col2',...))}

\item{delete}{delete action to be done as a boolean (TRUE/FALSE)}
}
\value{
ds same dataset as the inputed one but with survey marked for deletion if errors are found and delete=TRUE

errors  list of the errors found
}
\description{
This function check that all interviews in the dataset start after the actual first day of data collection.
There is an option to automatically mark for deletion the surveys which have started before the first day of data collection.
}
\examples{
\dontrun{
df <- sample_dataset
dates <- -c("survey_start","end_survey")
sc <- "survey_consent"
rc <- c("enumerator_id","X_uuid")
dl <- FALSE

chk3c_date_mistake(df, uuid, sc, rc, dl)
}
}
\author{
Yannick Pascaud
}
