% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFrequencyChecksSection4.R
\name{chk4biv_others_values}
\alias{chk4biv_others_values}
\title{List of other distinct values (not missing) per fields other with count}
\usage{
chk4biv_others_values(
  ds = NULL,
  otherpattern = NULL,
  enumeratorID = NULL,
  enumeratorcheck = FALSE
)
}
\arguments{
\item{ds}{dataset as a data.frame object}

\item{otherpattern}{pattern as string to identify the fields containing others values ('_other$')}

\item{enumeratorID}{name as a string of the field in the dataset where the enumerator ID is stored}

\item{enumeratorcheck}{specify if the report has to be displayed for each enumerator or not as a boolean (TRUE/FALSE)}
}
\value{
logf  the report
}
\description{
This function provide a report showing all distinct other values and the number of occurrences for each fields "other".
This report can be global (all the surveys) or displayed for each enumerator ID
}
\examples{
\dontrun{
df <- HighFrequencyChecks::sample_dataset
otherpattern <- "_other$"
enumeratorID <- "enumerator_id"
enumeratorcheck <- FALSE

log <- chk4biv_others_values(df, otherpattern, enumeratorID, enumeratorcheck)
head(log,10)
}
}
\author{
Yannick Pascaud
}
