% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFrequencyChecksSection5.R
\name{chk5d_duration_outliers}
\alias{chk5d_duration_outliers}
\title{Report the outlier durations for the surveys}
\usage{
chk5d_duration_outliers(
  ds = NULL,
  dates = NULL,
  sdval = NULL,
  reportingcol = NULL
)
}
\arguments{
\item{ds}{dataset as a data.frame object}

\item{dates}{fields as a list of string where the survey start and end date is stored (c('start_date','end_date'))}

\item{sdval}{number of standard deviation for which the duration within is considered as acceptable}

\item{reportingcol}{columns as a list of string name from the dataset you want in the result (c('col1','col2',...))}
}
\value{
logf  the report
}
\description{
This function report the outlier durations for the surveys
}
\examples{
\dontrun{
df <- sample_dataset
dt <- c("survey_start","end_survey")

chk5d_duration_outliers(df, dt)
}
}
\author{
Yannick Pascaud
}
