% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFrequencyChecksSection6.R
\name{chk6a_refusal}
\alias{chk6a_refusal}
\title{Check the percentage of survey refusals by enumerator}
\usage{
chk6a_refusal(ds = NULL, survey_consent = NULL, enumeratorID = NULL)
}
\arguments{
\item{ds}{dataset as a data.frame object}

\item{survey_consent}{name as a string of the field in the dataset where the survey consent is stored}

\item{enumeratorID}{name as a string of the field in the dataset where the enumerator ID is stored}
}
\value{
logf  the report
}
\description{
This function display the percentage of survey refusal per enumerator.
}
\examples{
\dontrun{
ds <- HighFrequencyChecks::sample_dataset
sc <- "survey_consent"
enumeratorID <- "enumerator_id"

log <- chk6a_refusal(ds, sc, enumeratorID)
head(log,10)
}
}
\author{
Yannick Pascaud
}
