% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFrequencyChecksSection6.R
\name{chk6b_duration}
\alias{chk6b_duration}
\title{Check the average interview duration by enumerator}
\usage{
chk6b_duration(ds = NULL, dates = NULL, enumeratorID = NULL)
}
\arguments{
\item{ds}{dataset as a data.frame object}

\item{dates}{fields as a list of string where the survey start and end date is stored (c('start_date','end_date'))}

\item{enumeratorID}{name as a string of the field in the dataset where the enumerator ID is stored}
}
\value{
logf  the report
}
\description{
This function display the average interview duration per enumerator.
}
\examples{
\dontrun{
ds <- HighFrequencyChecks::sample_dataset
dt <- c("survey_start","end_survey")
enumeratorID <- "enumerator_id"

log <- chk6b_duration(ds, dt, enumeratorID)
head(log,10)
}
}
\author{
Yannick Pascaud
}
