% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFrequencyChecksSection6.R
\name{chk6f_productivity}
\alias{chk6f_productivity}
\title{Check the surveyors with very low or high productivity}
\usage{
chk6f_productivity(
  ds = NULL,
  enumeratorID = NULL,
  surveydate = NULL,
  sdval = NULL
)
}
\arguments{
\item{ds}{dataset as a data.frame object}

\item{enumeratorID}{name as a string of the field in the dataset where the enumerator ID is stored}

\item{surveydate}{name as a string of the field in the dataset where the date of the survey is stored}

\item{sdval}{number of standard deviation for which the duration within is considered as acceptable}
}
\value{
logf  the report
}
\description{
This function display the surveyors with very low or high productivity.
}
\examples{
\dontrun{
ds <- HighFrequencyChecks::sample_dataset
enumeratorID <- "enumerator_id"
surveydate <- "survey_date"
sdval<-2

log <- chk6f_productivity(ds, enumeratorID, surveydate, sdval)
head(log,10)
}
}
\author{
Yannick Pascaud
}
