% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFrequencyChecksSection7.R
\name{chk7aii_productivity_hist}
\alias{chk7aii_productivity_hist}
\title{Overall productivity histogram}
\usage{
chk7aii_productivity_hist(
  ds = NULL,
  surveydate = NULL,
  dateformat = NULL,
  survey_consent = NULL
)
}
\arguments{
\item{ds}{dataset as a data.frame object}

\item{surveydate}{name as a string of the field in the dataset where the date of the survey is stored}

\item{dateformat}{format as a string used for the date ('\%m/\%d/\%Y')}

\item{survey_consent}{name as a string of the field in the dataset where the survey consent is stored}
}
\value{
graph  graphic as a plot.ly object
}
\description{
This function create an histogram showing the overall productivity per consent status per day.
}
\examples{
\dontrun{
df <- sample_dataset
sdte <- \"survey_date\"
dtf <- \"\\%m/\\%d/\\%Y\"
sc <- \"survey_consent\"

chk7aii_productivity_hist(df, sdte, dtf, sc)
}
}
\author{
Yannick Pascaud
}
