% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{enumeratorIsLazy}
\alias{enumeratorIsLazy}
\title{Check the enumerators who pick up less than X answers per specific question}
\usage{
enumeratorIsLazy(
  ds = NULL,
  enumeratorID = NULL,
  questions = NULL,
  minnbanswers = NULL
)
}
\arguments{
\item{ds}{dataset as a data.frame object}

\item{enumeratorID}{name as a string of the field in the dataset where the enumerator ID is stored}

\item{questions}{columns as a list of string name from the dataset you want to check against (c('col1','col2',...))}

\item{minnbanswers}{minimum number of answers expected per question}
}
\value{
logf  the report
}
\description{
This function display the surveyors who picked up less than a specified amount of answers per specific question.
This can be useful for select_multiple questions where respondent shall give at least 3 options for instance.
}
\examples{
\dontrun{
ds <- HighFrequencyChecks::sample_dataset
enumeratorID <- "enumerator_id"
questions <- c("consent_received.shelter_nfi.non_food_items[.]",
      "consent_received.food_security.main_income[.]",
      "consent_received.child_protection.boy_risk[.]",
      "consent_received.child_protection.girl_risk[.]")
mna <- 3

log <- enumeratorIsLazy(ds, enumeratorID, questions, mna)
head(log,10)
}
}
\author{
Yannick Pascaud
}
