% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{enumeratorProductivity}
\alias{enumeratorProductivity}
\title{Check the number of surveys by enumerator}
\usage{
enumeratorProductivity(ds = NULL, surveydate = NULL, enumeratorID = NULL)
}
\arguments{
\item{ds}{dataset as a data.frame object}

\item{surveydate}{name as a string of the field in the dataset where the date of the survey is stored}

\item{enumeratorID}{name as a string of the field in the dataset where the enumerator ID is stored}
}
\value{
logf  the report
}
\description{
This function display the total number of survey made and the average per day per enumerator.
}
\examples{
\dontrun{
ds <- HighFrequencyChecks::sample_dataset
surveydate <- "survey_date"
enumeratorID <- "enumerator_id"

log <- enumeratorProductivity(ds, surveydate, enumeratorID)
head(log,10)
}
}
\author{
Yannick Pascaud
}
