% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_e.R
\name{enumeratorProductivityOutliers}
\alias{enumeratorProductivityOutliers}
\title{Check the surveyors with very low or high productivity}
\usage{
enumeratorProductivityOutliers(
  ds = NULL,
  enumeratorID = NULL,
  surveyDate = NULL,
  sdval = 2
)
}
\arguments{
\item{ds}{dataset containing the survey (from kobo): data.frame}

\item{enumeratorID}{name of the field where the enumerator ID is stored: string}

\item{surveyDate}{name of the field in the dataset where the date of the survey is stored: string}

\item{sdval}{(Optional, by default set to 2) number of standard deviation for which the data within is considered as acceptable: integer}
}
\value{
dst      same dataset as the inputed one but with survey marked for deletion if errors are found and delete=TRUE (or NULL)

ret_log  list of the errors found (or NULL)

var      a list of value (or NULL)

graph    graphical representation of the results (or NULL)
}
\description{
This function display the surveyors with very low or high productivity.
}
\examples{
{
ds <- HighFrequencyChecks::sample_dataset
enumeratorID <- "enumerator_id"
surveyDate <- "survey_date"
sdval<-2

list[dst,ret_log,var,graph] <- enumeratorProductivityOutliers(ds=ds,
                                                              enumeratorID=enumeratorID,
                                                              surveyDate=surveyDate,
                                                              sdval=sdval)
head(ret_log,10)
print(graph)
}
}
\author{
Yannick Pascaud
}
