% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{enumeratorProductivityOutliers}
\alias{enumeratorProductivityOutliers}
\title{Check the surveyors with very low or high productivity}
\usage{
enumeratorProductivityOutliers(
  ds = NULL,
  enumeratorID = NULL,
  surveydate = NULL,
  sdval = NULL
)
}
\arguments{
\item{ds}{dataset as a data.frame object}

\item{enumeratorID}{name as a string of the field in the dataset where the enumerator ID is stored}

\item{surveydate}{name as a string of the field in the dataset where the date of the survey is stored}

\item{sdval}{number of standard deviation for which the duration within is considered as acceptable}
}
\value{
ret_log  the report
}
\description{
This function display the surveyors with very low or high productivity.
}
\examples{
{
ds <- HighFrequencyChecks::sample_dataset
enumeratorID <- "enumerator_id"
surveydate <- "survey_date"
sdval<-2

list[dst,ret_log,var,graph] <- enumeratorProductivityOutliers(ds,
                                                              enumeratorID,
                                                              surveydate,
                                                              sdval)
head(ret_log,10)
print(graph)
}
}
\author{
Yannick Pascaud
}
