% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_e.R
\name{enumeratorSurveysConsent}
\alias{enumeratorSurveysConsent}
\title{Check the percentage of survey refusals by enumerator}
\usage{
enumeratorSurveysConsent(ds = NULL, surveyConsent = NULL, enumeratorID = NULL)
}
\arguments{
\item{ds}{dataset containing the survey (from kobo): data.frame}

\item{surveyConsent}{name of the field in the dataset where the survey consent is stored: string}

\item{enumeratorID}{name of the field where the enumerator ID is stored: string}
}
\value{
dst      same dataset as the inputed one but with survey marked for deletion if errors are found and delete=TRUE (or NULL)

ret_log  list of the errors found (or NULL)

var      a list of value (or NULL)

graph    graphical representation of the results (or NULL)
}
\description{
This function display the percentage of survey refusal per enumerator.
}
\examples{
{
ds <- HighFrequencyChecks::sample_dataset
surveyConsent <- "survey_consent"
enumeratorID <- "enumerator_id"

list[dst,ret_log,var,graph] <- enumeratorSurveysConsent(ds=ds,
                                                        surveyConsent=surveyConsent,
                                                        enumeratorID=enumeratorID)
head(ret_log,10)
print(graph)
}
}
\author{
Yannick Pascaud
}
