% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{enumeratorSurveysConsent}
\alias{enumeratorSurveysConsent}
\title{Check the percentage of survey refusals by enumerator}
\usage{
enumeratorSurveysConsent(ds = NULL, survey_consent = NULL, enumeratorID = NULL)
}
\arguments{
\item{ds}{dataset as a data.frame object}

\item{survey_consent}{name as a string of the field in the dataset where the survey consent is stored}

\item{enumeratorID}{name as a string of the field in the dataset where the enumerator ID is stored}
}
\value{
ret_log  the report
}
\description{
This function display the percentage of survey refusal per enumerator.
}
\examples{
{
ds <- HighFrequencyChecks::sample_dataset
survey_consent <- "survey_consent"
enumeratorID <- "enumerator_id"

list[dst,ret_log,var,graph] <- enumeratorSurveysConsent(ds,
                                                        survey_consent,
                                                        enumeratorID)
head(ret_log,10)
}
}
\author{
Yannick Pascaud
}
