% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_i.R
\name{isInterviewAtTheSamplePoint}
\alias{isInterviewAtTheSamplePoint}
\title{GIS check surveys if fall without Xm radius from a sampled point}
\usage{
isInterviewAtTheSamplePoint(
  ds = NULL,
  dsCoordinates = NULL,
  sampledPoints = NULL,
  buffer = 10,
  surveyConsent = NULL,
  reportingColumns = c(enumeratorID, uniqueID),
  deleteIsInterviewAtTheSamplePoint = FALSE
)
}
\arguments{
\item{ds}{dataset containing the survey (from kobo): data.frame}

\item{dsCoordinates}{name of the fields from the dataset where the information about the GPS coordinates are stored: list of string (c('Long','Lat'))}

\item{sampledPoints}{dataset containing the shapefile of the households sampled - Regardless the projection used for the shapefile, it is transformed to WGS84}

\item{buffer}{value in meter to determine the buffer from a sampled point which is acceptable: integer}

\item{surveyConsent}{name of the field in the dataset where the survey consent is stored: string}

\item{reportingColumns}{(Optional, by default it is built from the enumeratorID and the UniqueID) name of the columns from the dataset you want in the result: list of string (c('col1','col2',...))}

\item{deleteIsInterviewAtTheSamplePoint}{(Optional, by default set as FALSE) if TRUE, the survey in error will be marked as 'deletedIsInterviewAtTheSamplePoint': boolean (TRUE/FALSE)}

\item{uniqueID}{name of the field where the survey unique ID is stored: string}

\item{enumeratorID}{name of the field where the enumerator ID is stored: string}
}
\value{
dst      same dataset as the inputed one but with survey marked for deletion if errors are found and delete=TRUE (or NULL)

ret_log  list of the errors found (or NULL)

var      a list of value (or NULL)

graph    graphical representation of the results (or NULL)
}
\description{
This function check that all interviews in the dataset were made within a distance from a sampled point.
It is based on a GIS shapefile providing the sample points for the assessment.
The function is based on the GPS data filled in the survey to determine their location.
There is an option to automatically mark for deletion the surveys which are to far away from a sampled point.

One internal function "make_GeodesicBuffer" used to create the buffers is created by Valentin
https://stackoverflow.com/users/5193830/valentin
}
\examples{
{
ds <- HighFrequencyChecks::sample_dataset
sampledPoints <- HighFrequencyChecks::SamplePts
dsCoordinates <- c("X_gps_reading_longitude","X_gps_reading_latitude")
buffer <- 10
surveyConsent <- "survey_consent"
uniqueID <- "X_uuid"
enumeratorID <- "enumerator_id"
reportingColumns <- c(enumeratorID, uniqueID)

list[dst,ret_log,var,graph] <- isInterviewAtTheSamplePoint(ds=ds,
                                                           dsCoordinates=dsCoordinates,
                                                           sampledPoints=sampledPoints,
                                                           buffer=buffer,
                                                           surveyConsent=surveyConsent,
                                                           reportingColumns=reportingColumns,
                                                           deleteIsInterviewAtTheSamplePoint=FALSE)
head(ret_log, 10)
}
}
\author{
Yannick Pascaud
}
