% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{isInterviewInTheCorrectSite}
\alias{isInterviewInTheCorrectSite}
\title{GIS check surveys for site}
\usage{
isInterviewInTheCorrectSite(
  adm = NULL,
  ds = NULL,
  ds_site = NULL,
  ds_coord = NULL,
  adm_site = NULL,
  survey_consent = NULL,
  reportingcol = NULL,
  correct = NULL
)
}
\arguments{
\item{adm}{dataset containing the shapefile - Regardless the projection used for the shapefile, it is transformed to WGS84}

\item{ds}{dataset as a data.frame object}

\item{ds_site}{name as a string of the field in the dataset where the site is stored}

\item{ds_coord}{columns as a list of string name from the dataset where the GPS coordinates are stored (c('Long','Lat'))}

\item{adm_site}{name as a string of the field in the shapefile where the site is stored}

\item{survey_consent}{name as a string of the field in the dataset where the survey consent is stored}

\item{reportingcol}{columns as a list of string name from the dataset you want in the result (c('col1','col2',...))}

\item{correct}{correction action to be done as a boolean (TRUE/FALSE)}
}
\value{
ds same dataset as the inputed one but with survey marked for deletion if errors are found and delete=TRUE

errors  list of the errors found
}
\description{
This function check that all interviews in the dataset were made in the correct site.
It is based on a GIS shapefile providing the boundaries of each site with their names.
The function is based on the GPS data filled in the survey to determine their location.
There is an option to automatically correct the site in the surveys whith the correct location.
}
\examples{
{
  admin <- HighFrequencyChecks::admin
  df <- HighFrequencyChecks::sample_dataset
  df_site <- "union_name"
  df_coord <- c("X_gps_reading_longitude","X_gps_reading_latitude")
  admin_site <- "Union"
  sc <- "survey_consent"
  reportingcol <- c("enumerator_id","X_uuid")
  correct <- FALSE

  list_site <- isInterviewInTheCorrectSite(admin, df, df_site, df_coord, admin_site, sc, reportingcol, correct)
  head(list_site[[2]], 10)
}
}
\author{
Yannick Pascaud
}
