% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{isInterviewTooShort}
\alias{isInterviewTooShort}
\title{Check that the duration of each interview is more than a threshold}
\usage{
isInterviewTooShort(
  ds = NULL,
  survey_consent = NULL,
  dates = NULL,
  reportingcol = NULL,
  minduration = 30,
  delete = NULL
)
}
\arguments{
\item{ds}{dataset as a data.frame object}

\item{survey_consent}{name as a string of the field in the dataset where the survey consent is stored}

\item{dates}{fields as a list of string where the survey start and end date is stored (c('start_date','end_date'))}

\item{reportingcol}{columns as a list of string name from the dataset you want in the result (c('col1','col2',...))}

\item{minduration}{minimum acceptable survey duration as integer in minutes}

\item{delete}{delete action to be done as a boolean (TRUE/FALSE)}
}
\value{
ds same dataset as the inputed one but with survey marked for deletion if errors are found and delete=TRUE

errors  list of the errors found
}
\description{
This function check that the duration of each interview is more than a specified threshold.
There is an option to automatically mark for deletion the surveys which are under the threshold.
Warning: If there are uncorrected mistakes in the survey dates, it can lead to have the length of the survey in seconds and this check will not performed well
}
\examples{
{
ds <- HighFrequencyChecks::sample_dataset
survey_consent <- "survey_consent"
dates <- c("survey_start","end_survey")
reportingcol <- c("enumerator_id","X_uuid")
minduration <- 30
delete <- FALSE

list <- isInterviewTooShort(ds, survey_consent, dates,  reportingcol, minduration, delete)
head(list[[2]], 10)
}
}
\author{
Yannick Pascaud
}
