% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{isSurveyStartedBeforeTheAssessment}
\alias{isSurveyStartedBeforeTheAssessment}
\title{Surveys that show start date earlier than first day of data collection}
\usage{
isSurveyStartedBeforeTheAssessment(
  ds = NULL,
  dates = NULL,
  survey_consent = NULL,
  start_collection = NULL,
  reportingcol = NULL,
  delete = NULL
)
}
\arguments{
\item{ds}{dataset as a data.frame object}

\item{dates}{fields as a list of string where the survey start and end date is stored (c('start_date','end_date'))}

\item{survey_consent}{name as a string of the field in the dataset where the survey consent is stored}

\item{start_collection}{date as a string of the first day of data collection ('yyyy-mm-dd')}

\item{reportingcol}{columns as a list of string name from the dataset you want in the result (c('col1','col2',...))}

\item{delete}{delete action to be done as a boolean (TRUE/FALSE)}
}
\value{
ds same dataset as the inputed one but with survey marked for deletion if errors are found and delete=TRUE

errors  list of the errors found
}
\description{
This function check that all interviews in the dataset start after the actual first day of data collection.
There is an option to automatically mark for deletion the surveys which have started before the first day of data collection.
}
\examples{
{
ds <- HighFrequencyChecks::sample_dataset
dates <- c("survey_start","end_survey")
survey_consent <- "survey_consent"
start_collection <- "2018-11-11"
reportingcol <- c("enumerator_id","X_uuid")
delete <- FALSE


list_date_mistake3 <- isSurveyStartedBeforeTheAssessment(ds, dates, survey_consent,start_collection, reportingcol, delete)
head(list_date_mistake3[[2]], 10)
}
}
\author{
Yannick Pascaud
}
