% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{isUniqueIDDuplicated}
\alias{isUniqueIDDuplicated}
\title{Duplicates in unique ID}
\usage{
isUniqueIDDuplicated(
  ds = NULL,
  UniqueID = NULL,
  survey_consent = NULL,
  reportingcol = NULL,
  delete = NULL
)
}
\arguments{
\item{ds}{dataset as a data.frame object}

\item{UniqueID}{name as a string of the field in the dataset where the unique ID is stored}

\item{survey_consent}{name as a string of the field in the dataset where the survey consent is stored}

\item{reportingcol}{columns as a list of string name from the dataset you want in the result (c('col1','col2',...))}

\item{delete}{delete action to be done as a boolean (TRUE/FALSE)}
}
\value{
dst same dataset as the inputed one but with survey marked for deletion if errors are found and delete=TRUE

ret_log  list of the errors found
}
\description{
This function check that all interviews in the dataset have an ID which is unique.
There is an option to automatically mark for deletion the surveys which have a duplicated unique ID.
}
\examples{
{
ds <- HighFrequencyChecks::sample_dataset
UniqueID <- "X_uuid"
survey_consent <- "survey_consent"
reportingcol <- c("enumerator_id","X_uuid")
delete <- FALSE


list[dst,ret_log,var,graph] <- isUniqueIDDuplicated(ds,
                                                    UniqueID,
                                                    survey_consent,
                                                    reportingcol,
                                                    delete)
head(ret_log, 10)
}
}
\author{
Yannick Pascaud
}
