% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{piechart}
\alias{piechart}
\title{Create a pie chart with ggplot}
\usage{
piechart(data, mapping)
}
\arguments{
\item{data}{dataset as a data.frame object}

\item{mapping}{aes paramters}
}
\value{
a ggplot object
}
\description{
This function allow to create a pie chart with ggplot.
}
\author{
https://ggplot2-book.org/programming.html
}
