% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportWrapper.R
\name{ReportWrapper}
\alias{ReportWrapper}
\title{Create an Rmd file for the report}
\usage{
ReportWrapper(
  ds = NULL,
  sdval = NULL,
  reportingcol = NULL,
  enumeratorID = NULL,
  enumeratorcheck = FALSE
)
}
\arguments{
\item{ds}{dataset as a data.frame object}

\item{sdval}{number of standard deviation for which the data within is considered as acceptable}

\item{reportingcol}{columns as a list of string name from the dataset you want in the result (c('col1','col2',...))}

\item{enumeratorID}{name as a string of the field in the dataset where the enumerator ID is stored}

\item{enumeratorcheck}{specify if the report has to be displayed for each enumerator or not as a boolean (TRUE/FALSE)}
}
\value{
vignette file
}
\description{
Create an Rmd file for the report
}
\examples{
\dontrun{
df <- HighFrequencyChecks::sample_dataset
sdval <- 2
reportingcol <- c("enumerator_id","X_uuid")
enumeratorID <- "enumerator_id"
enumeratorcheck <- FALSE

ReportWrapper(df, sdval, reportingcol, enumeratorID, enumeratorcheck)
}
}
\author{
Yannick Pascaud
}
