% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SamplePts}
\alias{SamplePts}
\title{points sampled to be surveyed dataset to be used as an example}
\format{
A Spatial Point data frame with 498 rows and 22 variables:
\describe{
  \item{\code{field_1}}{character COLUMN_DESCRIPTION}
  \item{\code{x}}{double COLUMN_DESCRIPTION}
  \item{\code{y}}{double COLUMN_DESCRIPTION}
  \item{\code{OBJECTID}}{character COLUMN_DESCRIPTION}
  \item{\code{OBJECTID_1}}{character COLUMN_DESCRIPTION}
  \item{\code{GEOCODE11}}{character COLUMN_DESCRIPTION}
  \item{\code{UpzCode}}{character COLUMN_DESCRIPTION}
  \item{\code{A4remark}}{character COLUMN_DESCRIPTION}
  \item{\code{adm4_en}}{character COLUMN_DESCRIPTION}
  \item{\code{adm3_en}}{character COLUMN_DESCRIPTION}
  \item{\code{adm2_en}}{character COLUMN_DESCRIPTION}
  \item{\code{adm1_en}}{character COLUMN_DESCRIPTION}
  \item{\code{adm0_en}}{character COLUMN_DESCRIPTION}
  \item{\code{a4code}}{character COLUMN_DESCRIPTION}
  \item{\code{a4code_old}}{character COLUMN_DESCRIPTION}
  \item{\code{a3code}}{character COLUMN_DESCRIPTION}
  \item{\code{a3code_old}}{character COLUMN_DESCRIPTION}
  \item{\code{a2code}}{character COLUMN_DESCRIPTION}
  \item{\code{a2code_old}}{character COLUMN_DESCRIPTION}
  \item{\code{a1code}}{character COLUMN_DESCRIPTION}
  \item{\code{a1code_old}}{character COLUMN_DESCRIPTION}
  \item{\code{a0code}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
SamplePts
}
\description{
This dataset is a sample of the points to be surveyed in both Ukhia and Teknaf.
For privacy purpose, the GPS coordinates are faked but are still located in the designated areas.
}
\keyword{internal}
