% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SampleSize}
\alias{SampleSize}
\title{sample size used}
\format{
A data frame with 11 rows and 7 variables:
\describe{
  \item{\code{Upazila}}{character COLUMN_DESCRIPTION}
  \item{\code{Union}}{character COLUMN_DESCRIPTION}
  \item{\code{NbHH}}{integer COLUMN_DESCRIPTION}
  \item{\code{Population}}{integer COLUMN_DESCRIPTION}
  \item{\code{EstPersPerShelter}}{double COLUMN_DESCRIPTION}
  \item{\code{SS}}{integer COLUMN_DESCRIPTION}
  \item{\code{TotPts}}{integer COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
SampleSize
}
\description{
This dataset comes from the Host Community Multi Sector Needs Assessment conducted in 2018 by IMPACT/REACH on behalf
of the ISCG and funded by ECHO. This is the sample size per union used in both Ukhia and Teknaf.
}
\keyword{internal}
