% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{surveyDistinctValues}
\alias{surveyDistinctValues}
\title{Number of distinct values (not missing) per fields}
\usage{
surveyDistinctValues(ds = NULL, enumeratorID = NULL, enumeratorcheck = FALSE)
}
\arguments{
\item{ds}{dataset as a data.frame object}

\item{enumeratorID}{name as a string of the field in the dataset where the enumerator ID is stored}

\item{enumeratorcheck}{specify if the report has to be displayed for each enumerator or not as a boolean (TRUE/FALSE)}
}
\value{
logf  the report
}
\description{
This function provide a report showing the number of distinct values for each fields.
This report can be global (all the surveys) or displayed for each enumerator ID
}
\examples{
\dontrun{
df <- HighFrequencyChecks::sample_dataset
enumeratorID <- "enumerator_id"
enumeratorcheck <- FALSE

log <- surveyDistinctValues(df, enumeratorID, enumeratorcheck)
head(log,10)
}
}
\author{
Yannick Pascaud
}
