% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{surveyOutliers}
\alias{surveyOutliers}
\title{Report the outlier values for all numerical field}
\usage{
surveyOutliers(
  ds = NULL,
  sdval = NULL,
  reportingcol = NULL,
  enumeratorID = NULL,
  enumeratorcheck = FALSE
)
}
\arguments{
\item{ds}{dataset as a data.frame object}

\item{sdval}{number of standard deviation for which the data within is considered as acceptable}

\item{reportingcol}{columns as a list of string name from the dataset you want in the result (c('col1','col2',...))}

\item{enumeratorID}{name as a string of the field in the dataset where the enumerator ID is stored}

\item{enumeratorcheck}{specify if the report has to be displayed for each enumerator or not as a boolean (TRUE/FALSE)}
}
\value{
logf  the report
}
\description{
This function provide a report showing all outlier values for each numerical fields.
The function will try to automatically determine the type of distribution (between Normal and Log-Normal)
based on the difference between mean and median between untransformed normalized and log transformed normalized distribution.
}
\examples{
\dontrun{
df <- HighFrequencyChecks::sample_dataset
sdval <- 2
reportingcol <- c("enumerator_id","X_uuid")
enumeratorID <- "enumerator_id"
enumeratorcheck <- FALSE

log <- surveyOutliers(df, sdval, reportingcol, enumeratorID, enumeratorcheck)
head(log,10)
}
}
\author{
Yannick Pascaud
}
